reducing opacity on adjacent tiles
