+++ src/pages/project/project.less
@@ -31,6 +31,10 @@
     left: 0;
     top: 0;
 
+    &.blurred {
+      opacity: 0.3;
+    }
+
     .block-interaction {
       position: absolute;
       top: 0;
